+++ src/Viewer/CoreGraph.cpp
@@ -935,7 +935,7 @@ Vwr::CoreGraph::CoreGraph( Data::Graph* graph, osg::ref_ptr<osg::Camera> camera 
 	manipulator_rotation->addTransformUpdating( graphRotTransf );
 
 	float scale = graphRotTransf->getBound().radius() * 2.0f;
-	osg::Matrix mat_rot = osg::Matrix::scale( scale, scale, scale ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
+	osg::Matrix mat_rot = osg::Matrix::scale( static_cast<double>(scale), static_cast<double>(scale), static_cast<double>(scale) ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
 
 	manipulator_scale->setMatrix( mat_rot );
 	manipulator_rotation->setMatrix( mat_rot );
@@ -1032,7 +1032,7 @@ void CoreGraph::toggleDragger( int dragger_no, bool set )
 			root->addChild( manipulator_rotation );
 			manipulator_rotation->addTransformUpdating( graphRotTransf );
 			float scale = graphRotTransf->getBound().radius() * 1.3f;
-			osg::Matrix mat_rot = osg::Matrix::scale( scale, scale, scale ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
+			osg::Matrix mat_rot = osg::Matrix::scale( static_cast<double>(scale), static_cast<double>(scale), static_cast<double>(scale) ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
 			manipulator_rotation->setMatrix( mat_rot );
 			manipulator_rotation->setHandleEvents( set );
 			manipulator_rotation->setNodeMask( set? 0xffffffff : 0x0 );
@@ -1154,7 +1154,7 @@ void CoreGraph::reload( Data::Graph* graph )
 	this->browsersGroup->getGroup()->getOrCreateStateSet()->setRenderBinDetails( 100,"RenderBin" );
 
 	float scale = graphRotTransf->getBound().radius() * 1.5f;
-	osg::Matrix mat = osg::Matrix::scale( scale, scale, scale ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
+	osg::Matrix mat = osg::Matrix::scale( static_cast<double>(scale), static_cast<double>(scale), static_cast<double>(scale) ) * osg::Matrix::translate( graphRotTransf->getBound().center() );
 	manipulator_scale->setMatrix( mat );
 	manipulator_rotation->setMatrix( mat );
 }
@@ -2026,10 +2026,10 @@ void CoreGraph::createBase()
 //set aruco modelView matrix
 void CoreGraph::recievedMVMatrix( QMatrix4x4 modelViewMatrix )
 {
-	osg::Matrixd arucoMVM( modelViewMatrix.operator()( 0,0 ),modelViewMatrix.operator()( 0,1 ),modelViewMatrix.operator()( 0,2 ),modelViewMatrix.operator()( 0,3 ),
-						   modelViewMatrix.operator()( 1,0 ),modelViewMatrix.operator()( 1,1 ),modelViewMatrix.operator()( 1,2 ),modelViewMatrix.operator()( 1,3 ),
-						   modelViewMatrix.operator()( 2,0 ),modelViewMatrix.operator()( 2,1 ),modelViewMatrix.operator()( 2,2 ),modelViewMatrix.operator()( 2,3 ),
-						   modelViewMatrix.operator()( 3,0 ),modelViewMatrix.operator()( 3,1 ),modelViewMatrix.operator()( 3,2 ),modelViewMatrix.operator()( 3,3 ) );
+	osg::Matrixd arucoMVM( static_cast<double>(modelViewMatrix.operator()( 0,0 )),static_cast<double>(modelViewMatrix.operator()( 0,1 )),static_cast<double>(modelViewMatrix.operator()( 0,2 )),static_cast<double>(modelViewMatrix.operator()( 0,3 )),
+						   static_cast<double>(modelViewMatrix.operator()( 1,0 )),static_cast<double>(modelViewMatrix.operator()( 1,1 )),static_cast<double>(modelViewMatrix.operator()( 1,2 )),static_cast<double>(modelViewMatrix.operator()( 1,3 )),
+						   static_cast<double>(modelViewMatrix.operator()( 2,0 )),static_cast<double>(modelViewMatrix.operator()( 2,1 )),static_cast<double>(modelViewMatrix.operator()( 2,2 )),static_cast<double>(modelViewMatrix.operator()( 2,3 )),
+						   static_cast<double>(modelViewMatrix.operator()( 3,0 )),static_cast<double>(modelViewMatrix.operator()( 3,1 )),static_cast<double>(modelViewMatrix.operator()( 3,2 )),static_cast<double>(modelViewMatrix.operator()( 3,3 )) );
 
 	camera->setViewMatrix( arucoMVM );
 	//update base size
@@ -2040,10 +2040,10 @@ void CoreGraph::recievedMVMatrix( QMatrix4x4 modelViewMatrix )
 //set aruco projection matrix
 void CoreGraph::recievedPMatrix( QMatrix4x4 projectionMatrix )
 {
-	osg::Matrixd arucoPM( projectionMatrix.operator()( 0,0 ),projectionMatrix.operator()( 0,1 ),projectionMatrix.operator()( 0,2 ),projectionMatrix.operator()( 0,3 ),
-						  projectionMatrix.operator()( 1,0 ),projectionMatrix.operator()( 1,1 ),projectionMatrix.operator()( 1,2 ),projectionMatrix.operator()( 1,3 ),
-						  projectionMatrix.operator()( 2,0 ),projectionMatrix.operator()( 2,1 ),projectionMatrix.operator()( 2,2 ),projectionMatrix.operator()( 2,3 ),
-						  projectionMatrix.operator()( 3,0 ),projectionMatrix.operator()( 3,1 ),projectionMatrix.operator()( 3,2 ),projectionMatrix.operator()( 3,3 ) );
+	osg::Matrixd arucoPM( static_cast<double>(projectionMatrix.operator()( 0,0 )),static_cast<double>(projectionMatrix.operator()( 0,1 )),static_cast<double>(projectionMatrix.operator()( 0,2 )),static_cast<double>(projectionMatrix.operator()( 0,3 )),
+						  static_cast<double>(projectionMatrix.operator()( 1,0 )),static_cast<double>(projectionMatrix.operator()( 1,1 )),static_cast<double>(projectionMatrix.operator()( 1,2 )),static_cast<double>(projectionMatrix.operator()( 1,3 )),
+						  static_cast<double>(projectionMatrix.operator()( 2,0 )),static_cast<double>(projectionMatrix.operator()( 2,1 )),static_cast<double>(projectionMatrix.operator()( 2,2 )),static_cast<double>(projectionMatrix.operator()( 2,3 )),
+						  static_cast<double>(projectionMatrix.operator()( 3,0 )),static_cast<double>(projectionMatrix.operator()( 3,1 )),static_cast<double>(projectionMatrix.operator()( 3,2 )),static_cast<double>(projectionMatrix.operator()( 3,3 )) );
 
 	camera->setProjectionMatrix( arucoPM );
 }
